+++ src/test/kotlin/cl/ravenhill/jakt/constraints/doubles/BeAtLeastTest.kt
@@ -1,12 +1,11 @@
 package cl.ravenhill.jakt.constraints.doubles
 
 import cl.ravenhill.jakt.arbs.datatypes.real
-import cl.ravenhill.jakt.testfactories.`validate BeAtLeastConstraint`
+import cl.ravenhill.jakt.testfactories.`validate BeAtLeast constraint`
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.matchers.shouldBe
 import io.kotest.property.Arb
 
 class BeAtLeastTest : FreeSpec({
 
-    include(`validate BeAtLeastConstraint`(Arb.real(), ::BeAtLeast))
+    include(`validate BeAtLeast constraint`(Arb.real(), ::BeAtLeast))
 })